Dataset File
This is the sample CSV file that i used for this code, You need to run a cell of mounting the csv with your ipynb file to perform execution of the code. for this purpose u may use Pandas library. 
import pandas as pd
DataFrame = pd.read_csv(r'filepath/heart.csv')
print(df)